+++ vsc-organize-imports/out/SortImports.js
@@ -10,115 +10,123 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const ts = require("typescript");
 const vsc = require("vsc-base");
-const vscode = require("vscode");
-function SortImports(content, spaceBetweenImportGroups, orderSpecifiers, orderSpecifiersAsSingleLine) {
+function SortImports(basePath, content, emptyLinesAfterGlobalImports, emptyLinesAfterAbsoluteImports, emptyLinesLocalImports, emptyLinesAfterImports, orderSpecifiers, orderSpecifiersAsSingleLine) {
     return __awaiter(this, void 0, void 0, function* () {
+        // Find first non imports: (exclude 'use strict' and sourceFile')
+        const sourceFile = vsc.tsCreateSourceFile(content);
+        const children = vsc.tsGetParsedChildren(sourceFile);
+        const firstNode = children.find(node => !ts.isImportDeclaration(node)
+            && !(ts.isExpressionStatement(node) && node.expression.getText() === 'use strict'));
+        const _imports = children.filter(node => ts.isImportDeclaration(node) && (!firstNode || node.pos < firstNode.pos));
         //Find first node that is not in import
-        const imports = mapImports(content);
+        const imports = mapImports(content, _imports);
         if (!imports) {
             return Promise.resolve(undefined);
         }
         //find last import
-        const firstImport = imports.sort((a, b) => a.pos.end - b.pos.end)[0];
+        const firstImport = imports[0];
         const lastImport = imports[imports.length - 1];
-        //sort
-        imports.sort((a, b) => a.path.localeCompare(b.path));
-        //TODO: sort specifiers (maybe)
         if (orderSpecifiers) {
-            imports.map(imp => {
-                if (imp.specifiers) {
-                    imp.specifiers.sort((a, b) => a.name.localeCompare(b.name));
-                    if (orderSpecifiersAsSingleLine) {
-                        const specifierContent = imp.specifiers.map(s => s.fullString).join(', ');
-                        imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{ ' + specifierContent + ' }');
-                    }
-                    else {
-                        const specifierContent = imp.specifiers.map(s => s.fullString).join(',\n  ');
-                        imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{\n  ' + specifierContent + '\n}');
-                    }
-                }
-            });
+            sortNamedImports(imports, orderSpecifiersAsSingleLine);
         }
-        const newImportContent = yield organizeImports(imports, spaceBetweenImportGroups);
-        //await vsc.insertAt(newImportContent, firstImport.pos.start, lastImport.pos.end)
-        //todo redo this and use insertAt (when vsc-base use the same reaplce)
-        const range = new vscode.Range(firstImport.pos.startPosition, lastImport.pos.endPosition);
-        const edits = [];
-        edits.push(vscode.TextEdit.replace(range, newImportContent));
-        const workspaceEdit = new vscode.WorkspaceEdit();
-        workspaceEdit.set(vscode.window.activeTextEditor.document.uri, edits);
-        vscode.workspace.applyEdit(workspaceEdit);
+        const newImportContent = yield organizeImports(imports, basePath, emptyLinesAfterGlobalImports, emptyLinesAfterAbsoluteImports, emptyLinesLocalImports, emptyLinesAfterImports);
+        let contentToFirst;
+        let end = lastImport.node.end;
+        end += content.substr(end).match(/[\n\s]*/)[0].length;
+        vsc.insertAt(newImportContent, firstImport.node.pos, end);
     });
 }
 exports.SortImports = SortImports;
-const organizeImports = (imports, spaceBetweenImportGroups) => __awaiter(this, void 0, void 0, function* () {
-    // load dependencies to specify global imports
-    const [dependencies, devDependencies] = yield vsc.getPackageDependencies();
-    const dependencyNames = Object.keys(Object.assign({}, dependencies, devDependencies));
-    //split into global / local
-    const globalImports = [];
-    const absoluteImports = [];
-    const localImports = [];
-    const directImports = [];
-    imports.forEach(_import => {
-        const global = !!dependencyNames.find(name => {
-            return name.indexOf(_import.path) === 0;
-        });
-        const local = /^\./.test(_import.path);
-        if (!_import.node.importClause) {
-            directImports.push(_import);
-        }
-        else if (global) {
-            globalImports.push(_import);
-        }
-        else if (local) {
-            localImports.push(_import);
-        }
-        else {
-            absoluteImports.push(_import);
-        }
-    });
-    const addNewLine = () => {
-        if (newImportContent.length > 0) {
-            newImportContent += '\n';
-            if (spaceBetweenImportGroups) {
-                newImportContent += '\n';
+const sortNamedImports = (imports, orderSpecifiersAsSingleLine) => {
+    imports.map(imp => {
+        if (imp.specifiers) {
+            imp.specifiers.sort((a, b) => a.name.localeCompare(b.name));
+            if (orderSpecifiersAsSingleLine) {
+                const specifierContent = imp.specifiers.map(s => s.fullString).join(', ');
+                imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{ ' + specifierContent + ' }');
+            }
+            else {
+                const specifierContent = imp.specifiers.map(s => s.fullString).join(',\n  ');
+                imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{\n  ' + specifierContent + '\n}');
             }
         }
+    });
+};
+const organizeImports = (imports, basePath, emptyLinesAfterGlobalImports, emptyLinesAfterAbsoluteImports, emptyLinesLocalImports, emptyLinesAfterImports) => __awaiter(this, void 0, void 0, function* () {
+    const moduleExtensionRegExp = /\.([tj]sx?)+$/;
+    const localRegExp = /^\./;
+    const groups = {
+        localImports: [],
+        absoluteImports: [],
+        directImports: [],
+        globalImports: [],
     };
-    //organize in groups with spaces between
+    //split into global / local
+    imports.forEach(_import => {
+        if (localRegExp.test(_import.path)) {
+            groups.localImports.push(_import);
+            return;
+        }
+        const fullPath = vsc.joinPaths(basePath, _import.path);
+        if (moduleExtensionRegExp.test(_import.path) && vsc.doesExists(fullPath)) {
+            groups.absoluteImports.push(_import);
+            return;
+        }
+        if (vsc.doesExists(fullPath + '.ts') ||
+            vsc.doesExists(fullPath + '.tsx') ||
+            vsc.doesExists(fullPath + '.js') ||
+            vsc.doesExists(fullPath + '.jsx')) {
+            groups.absoluteImports.push(_import);
+            return;
+        }
+        if (!_import.node.importClause) {
+            groups.directImports.push(_import);
+            return;
+        }
+        groups.globalImports.push(_import);
+    });
+    const defaultMapping = [
+        { groups: ['globalImports'], emptyLinesAfterGroup: emptyLinesAfterGlobalImports },
+        { groups: ['absoluteImports'], emptyLinesAfterGroup: emptyLinesAfterAbsoluteImports },
+        { groups: ['localImports'], emptyLinesAfterGroup: emptyLinesLocalImports },
+        { groups: ['directImports'], emptyLinesAfterGroup: 0 },
+    ];
     let newImportContent = "";
-    if (globalImports.length > 0) {
-        newImportContent += globalImports.map(imp => imp.fullString).join('\n');
-    }
-    if (absoluteImports.length > 0) {
-        addNewLine();
-        newImportContent += absoluteImports.map(imp => imp.fullString).join('\n');
-    }
-    if (localImports.length > 0) {
-        addNewLine();
-        newImportContent += localImports.map(imp => imp.fullString).join('\n');
-    }
-    if (directImports.length > 0) {
-        addNewLine();
-        newImportContent += directImports.map(imp => imp.fullString).join('\n');
+    defaultMapping.forEach((groupOptions, index) => {
+        let group = [];
+        groupOptions.groups.forEach(groupName => {
+            group = [...groups[groupName]];
+        });
+        if (group.length === 0) {
+            return;
+        }
+        // sort
+        group.sort((a, b) => a.path.localeCompare(b.path));
+        // join and add
+        newImportContent += group.map(imp => imp.fullString).join('\n') + '\n';
+        // add spaces
+        for (let space = 0; space < groupOptions.emptyLinesAfterGroup; space++) {
+            newImportContent += '\n';
+        }
+    });
+    newImportContent = newImportContent.trim() + '\n';
+    for (let lines = 0; lines < emptyLinesAfterImports; lines++) {
+        newImportContent += '\n';
     }
     return newImportContent;
 });
-const mapImports = (content) => {
-    const [, firstPos] = vsc.tsFindNodePositionFromContent(content, node => !ts.isSourceFile(node)
-        && !ts.isImportDeclaration(node)
-        && !(ts.isExpressionStatement(node) && node.expression.getText() === 'use strict'));
+const mapImports = (content, _imports) => {
     // All imports before first statement, mapped with import path
-    const importPos = vsc.tsFindAllNodePositionsFromContent(content, node => ts.isImportDeclaration(node) && (!firstPos || node.end < firstPos.start));
-    const imports = importPos.map(([node, pos], index) => {
+    // Map with name?, fullString, and named imports info
+    const imports = _imports.map((node, index) => {
         const importNode = node;
         let name = '';
         const fullString = content
-            .substring(index === 0 ? pos.start : importPos[index - 1][1].end + 1, pos.end)
+            .substring(index === 0 ? node.pos : _imports[index - 1].end + 1, node.end)
             .trim();
         let specifiers = [];
         const importClause = importNode.importClause;
+        //named imports (specifiers)
         if (importClause) {
             if (importClause.name) {
                 name = importClause.name.getText();
@@ -132,7 +140,6 @@ const mapImports = (content) => {
             }
         }
         return ({
-            pos,
             name,
             specifiers,
             fullString,