+++ vsc-organize-imports/src/OrganizeImports.ts
@@ -1,6 +1,7 @@
 'use strict'
 import * as vsc from 'vsc-base'
 import * as vscode from 'vscode'
+
 import { SortImports } from './SortImports'
 
 export default class OrganizeImports {
@@ -15,12 +16,34 @@ export default class OrganizeImports {
       if (!content) {
          return
       }
+      const path = uri.path
+
       //load settings:
-      const spaceBetweenImportGroups = this.getConfig('spaceBetweenImportGroups', false)
       const orderSpecifiers = this.getConfig('orderSpecifiers', false)
       const orderSpecifiersAsSingleLine = this.getConfig('orderSpecifiersAsSingleLine', false)
+      const baseUrl = this.getConfig('baseUrl', 'src')
 
-      await SortImports(content, spaceBetweenImportGroups, orderSpecifiers, orderSpecifiersAsSingleLine)
+      const emptyLinesAfterGlobalImports = this.getConfig('emptyLinesAfterGlobalImports', 0)
+      const emptyLinesAfterAbsoluteImports = this.getConfig('emptyLinesAfterAbsoluteImports', 0)
+      const emptyLinesLocalImports = this.getConfig('emptyLinesLocalImports', 0)
+      const emptyLinesAfterImports = this.getConfig('emptyLinesAfterImports', 1)
+
+      const rootPath = vsc.getRootPath(path)
+      if (!rootPath) {
+         return
+      }
+      const basePath = vsc.joinPaths(rootPath, baseUrl)
+
+      await SortImports(
+         basePath,
+         content,
+         emptyLinesAfterGlobalImports,
+         emptyLinesAfterAbsoluteImports,
+         emptyLinesLocalImports,
+         emptyLinesAfterImports,
+         orderSpecifiers,
+         orderSpecifiersAsSingleLine
+      )
    }
 }